add config validator
